Update stone.js
I don't use github. or code, so do check the changes, doubly so.
I made a request, but now found? where to do it myself, to add the S3 atk/atk stone (+4000, +37%), I use it on Bel and according to my testing it is the best for her, but i'd like to confirm using this magnificent tool of yours that you have so graciously provided to the community.
Sorry to bother you, and thank you.